Further implementation. Moved Driver to interfaces
